requirements.txt
pynsq
clime
tornado
ptt-crawler
